new frontend organization and structure, with vanilla js
