+++ cabraping.fe/src/components/footer/html.js
@@ -1,4 +1,5 @@
-const Footer = () => {
+// const Footer = () => {
+const Footer_html = () => {
   const view = `
     <div class="d-flex justify-content-center align-items-center p-3 bg-dark">
         <span class="text-white">CabraPing. All rights reserved © 2023</span>
@@ -8,4 +9,5 @@ const Footer = () => {
   return view;
 };
 
-export default Footer;
+// export default Footer;
+export default Footer_html;
