+++ cabraping.fe/src/components/header/funcions-js.js
@@ -1,0 +1,5 @@
+
+// export function HeaderInit() {
+export function Header_js() {
+  console.log("hello");
+}
