+++ cabraping.fe/src/components/showNotification.js
@@ -1,0 +1,14 @@
+
+export function showNotification(message, type) {
+  const container = document.getElementById('notification-container');
+  const notification = document.createElement('div');
+  notification.classList.add('notification', type);
+  notification.textContent = message;
+
+  container.appendChild(notification);
+
+  // Automatically remove the notification after 5 seconds
+  setTimeout(() => {
+      container.removeChild(notification);
+  }, 5000);
+}
