--- cabraping.fe/src/pages/Error404.js
@@ -1,13 +1,0 @@
-export function Error404Init() {
-
-  console.log("start code in Error404");
-}
-
-export function Error404() {
-  return `
-    <div class="Error404">
-      <h2>Error 404</h2>
-    </div>
-  `;
-};
-
