+++ cabraping.fe/src/pages/Error404/funcions-js.js
@@ -1,0 +1,5 @@
+
+export function Error404_js() {
+
+  console.log("start code in Error404");
+}
