+++ cabraping.fe/src/pages/Error404/html.js
@@ -1,0 +1,9 @@
+
+export function Error404_html() {
+  return `
+    <div class="Error404">
+      <h2>Error 404</h2>
+    </div>
+  `;
+};
+
