+++ cabraping.fe/src/pages/Game/html.js
@@ -1,0 +1,11 @@
+
+export function Game_html() {
+  return `
+    <div class="w-100 h-100 d-flex gap-4">
+      <h1>game</h1>
+      <div class="game-board">
+        <canvas width="750" height="585" id="game"></canvas>
+      </div>
+    </div>
+  `;
+}
