+++ cabraping.fe/src/pages/Logout/funcions-js.js
@@ -1,4 +1,4 @@
-export function LogoutPage() {
+export function LogoutPage_js() {
   localStorage.removeItem("jwt");
   window.location.replace("");
   return "";
