--- cabraping.fe/src/pages/User.js
@@ -1,11 +1,0 @@
-export function UserInit() {
-  console.log("Start code in User");
-}
-
-export function User() {
-  return `
-    <div class="User" >
-        <h2> User </h2>
-    </div>
-  `;
-};
