+++ cabraping.fe/src/pages/User/funcions-js.js
@@ -1,0 +1,4 @@
+// export function UserInit() {
+export function User_js() {
+  console.log("Start code in User");
+}
