+++ cabraping.fe/src/pages/User/html.js
@@ -1,0 +1,9 @@
+
+// export function User() {
+export function User_html() {
+  return `
+    <div class="User" >
+        <h2> User </h2>
+    </div>
+  `;
+};
