+++ cabraping.fe/src/pages/Users/html.js
@@ -1,0 +1,15 @@
+import { getToken } from "../../utils/get-token";
+
+// export function Users() {
+export function Users_html() {
+  const jwt = getToken();
+
+  return `
+    <div class="container-sm">
+      <h2>List of Users</h2>
+      <ul id="users-list" class="list-group">
+        <!-- userData -->
+      </ul>
+    </div>
+  `;
+}
