+++ cabraping.fe/src/routes/index.js
@@ -1,64 +1,56 @@
-// import Header from "../template/Header";
-import {Header, HeaderInit} from "../template/Header";
-import Footer from "../template/Footer";
-import { Home, HomeInit } from "../pages/Home";
-import { Game, GameInit } from "../pages/Game";
-import { Chat, ChatInit } from "../pages/Chat";
-import { User, UserInit } from "../pages/User";
-import { Users, UsersInit } from "../pages/Users";
-import {
-  Friends,
-  FriendsRender,
-  FriendRequestsRender,
-  FriendsInit,
-} from "../pages/Friends";
-import { AuthPage, AuthPageInit } from "../pages/AuthPage";
-import { LogoutPage } from "../pages/Logout";
-import { Error404, Error404Init } from "../pages/Error404";
-import getHash from "../utils/getHash";
+import Footer_html from "../components/footer/html";
+import { Home_js } from "../pages/home/funcions-js";
+import { Home_html } from "../pages/home/html";
+import { AuthPage_js } from "../pages/AuthPage/funcions-js";
+import { AuthPage_html } from "../pages/AuthPage/html";
+import { Error404_js } from "../pages/Error404/funcions-js";
+import { Error404_html } from "../pages/Error404/html";
+import { User_js } from "../pages/User/funcions-js";
+import { User_html } from "../pages/User/html";
+import { LogoutPage_js } from "../pages/Logout/funcions-js";
+import { Users_js } from "../pages/Users/funcions-js";
+import { Users_html } from "../pages/Users/html";
+import { Header_js} from "../components/header/funcions-js";
+import { Header_html } from "../components/header/html";
+import { Game_js } from "../pages/Game/funcions-js";
+import { Game_html } from "../pages/Game/html";
+import { Friends_html } from "../pages/Friends/html";
+import { FriendsRender, FriendRequestsRender, Friends_js } from "../pages/Friends/funcions-js";
+import { Chat_js } from "../pages/Chat/funcions-js";
+import { Chat_html } from "../pages/Chat/html";
+import resolveRoutes from "../utils/resolveRoutes";
 
 const routes = {
-  "/": [Home, HomeInit],
-  "/auth": [AuthPage, AuthPageInit],
-  "/logout": [LogoutPage],
-  "/game": [Game, GameInit],
-  "/chat": [Chat, ChatInit],
-  "/chat/:id": [Chat, ChatInit],
-  "/user": [User, UserInit],
-  "/users": [Users, UsersInit],
-  "/friends": [Friends, FriendsRender, FriendRequestsRender, FriendsInit],
+  "/": [Home_html, Home_js],
+  "/auth": [AuthPage_html, AuthPage_js],
+  "/logout": [LogoutPage_js],
+  "/game": [Game_html, Game_js],
+  "/chat": [Chat_html, Chat_js],
+  "/chat/:id": [Chat_html, Chat_js],
+  "/user": [User_html, User_js],
+  "/users": [Users_html, Users_js],
+  "/friends": [Friends_html, FriendsRender, FriendRequestsRender, Friends_js],
+  "/404": [Error404_html, Error404_js],
 };
 
 const router = async () => {
   // The variable where it is stored, the array of the functions that will be executed
   let render;
-  let catch_path;
 
   // The HTML component selectors where our components will be rendered
-  const header = null || document.getElementById("header");
-  const content = null || document.getElementById("content");
-  const footer = null || document.getElementById("footer");
+  const header = document.getElementById("header");
+  const footer = document.getElementById("footer");
+  const content = document.getElementById("content");
 
-  // We will render our Header and Footer components
-  header.innerHTML = await Header();
-  footer.innerHTML = await Footer();
-  HeaderInit();
+  footer.innerHTML = await Footer_html();
+  header.innerHTML = await Header_html();
+  Header_js();
 
   // We take our fruit and divide it into "/", we obtain a lowercase array, to work with the paths
   let user_location = location.hash.slice(1).toLocaleLowerCase().split("/");
 
-  // handling getHash, or route ids
-  if (getHash() != "/") catch_path = `/${user_location[0]}/:id`;
-  else catch_path = `/${user_location[0]}`;
-
   // We check that the route exists.
-  if (
-    user_location.length >= 1 &&
-    user_location.length <= 2 &&
-    routes[catch_path]
-  )
-    render = routes[`/${user_location[0]}`];
-  else render = [Error404, Error404Init];
+  render = resolveRoutes(routes, user_location);
 
   // We render our html and execute the js
   content.innerHTML = await render[0]();
